Close #39 add logo in qr code image
