+++ app/models/concerns/scorecard_progresses/scorecard_callback_concern.rb
@@ -7,8 +7,6 @@ module ScorecardProgresses::ScorecardCallbackConcern
     after_save :set_scorecard_progress
     after_save :update_counter_cache, if: :downloaded?
     after_destroy :update_counter_cache, if: :downloaded?
-    after_save :generate_qr_code, if: :open_voting?
-    after_save :calculate_voting_results, if: :close_voting?
 
     private
       def update_counter_cache
@@ -31,19 +29,6 @@ module ScorecardProgresses::ScorecardCallbackConcern
             "running_date" => conducted_at
           }
         }
-      end
-
-      def generate_qr_code
-        Scorecards::OpenVotingService.new(scorecard).call
-      end
-
-      def calculate_voting_results
-        scorecard.voting_indicators.includes(:ratings).find_each do |indicator|
-          scores = indicator.ratings.pluck(:score)
-          median = Ratings::MedianCalculator.new(scores).call
-
-          indicator.update_column(:median, median)
-        end
       end
   end
 end
