+++ spec/models/scorecard_progress_spec.rb
@@ -82,89 +82,6 @@ RSpec.describe ScorecardProgress, type: :model do
     end
   end
 
-  describe "#after_save: generate_qr_code" do
-    context "when status is open_voting" do
-      let!(:scorecard) { create(:scorecard, progress: :running) }
-
-      it "generates a QR code for the scorecard" do
-        expect(scorecard.qr_code.present?).to be_falsey
-
-        create(:scorecard_progress, status: :open_voting, scorecard: scorecard)
-
-        expect(scorecard.reload.qr_code.present?).to be_truthy
-      end
-    end
-
-    context "when status is not open_voting" do
-      let!(:scorecard) { create(:scorecard, progress: nil) }
-
-      it "does not generate a QR code" do
-        create(:scorecard_progress, status: :downloaded, scorecard: scorecard)
-
-        expect(scorecard.reload.qr_code.present?).to be_falsey
-      end
-    end
-  end
-
-  describe "#after_save: calculate_voting_results" do
-    context "when status is close_voting" do
-      let!(:scorecard) { create(:scorecard, progress: :open_voting) }
-      let!(:scorecard_progress) { build(:scorecard_progress, scorecard: scorecard, status: :close_voting) }
-
-      it "calculates voting results for all indicators" do
-        indicator1 = create(:voting_indicator, scorecard: scorecard, median: nil)
-        indicator2 = create(:voting_indicator, scorecard: scorecard, median: nil)
-
-        create(:rating, voting_indicator_uuid: indicator1.uuid, scorecard_uuid: scorecard.uuid, score: 1)
-        create(:rating, voting_indicator_uuid: indicator1.uuid, scorecard_uuid: scorecard.uuid, score: 3)
-        create(:rating, voting_indicator_uuid: indicator1.uuid, scorecard_uuid: scorecard.uuid, score: 5)
-
-        create(:rating, voting_indicator_uuid: indicator2.uuid, scorecard_uuid: scorecard.uuid, score: 2)
-        create(:rating, voting_indicator_uuid: indicator2.uuid, scorecard_uuid: scorecard.uuid, score: 2)
-        create(:rating, voting_indicator_uuid: indicator2.uuid, scorecard_uuid: scorecard.uuid, score: 4)
-
-        scorecard_progress.save
-
-        expect(indicator1.reload.median_before_type_cast).to eq(3)
-        expect(indicator2.reload.median_before_type_cast).to eq(2)
-      end
-
-      it "sets median to nil when there are no ratings" do
-        indicator = create(:voting_indicator, scorecard: scorecard, median: 1)
-
-        scorecard_progress.save
-
-        expect(indicator.reload.median).to be_nil
-      end
-
-      it "ignores nil scores when calculating median" do
-        indicator = create(:voting_indicator, scorecard: scorecard, median: nil)
-
-        create(:rating, voting_indicator_uuid: indicator.uuid, scorecard_uuid: scorecard.uuid, score: nil)
-        create(:rating, voting_indicator_uuid: indicator.uuid, scorecard_uuid: scorecard.uuid, score: 1)
-        create(:rating, voting_indicator_uuid: indicator.uuid, scorecard_uuid: scorecard.uuid, score: 3)
-
-        scorecard_progress.save
-
-        expect(indicator.reload.median_before_type_cast).to eq(2)
-      end
-    end
-
-    context "when status is not close_voting" do
-      let!(:scorecard) { create(:scorecard, progress: :running) }
-
-      it "does not calculate voting results" do
-        indicator = create(:voting_indicator, scorecard: scorecard, median: nil)
-
-        create(:rating, voting_indicator_uuid: indicator.uuid, scorecard_uuid: scorecard.uuid, score: 5)
-
-        create(:scorecard_progress, status: :open_voting, scorecard: scorecard)
-
-        expect(indicator.reload.median).to be_nil
-      end
-    end
-  end
-
   describe "#after_save update counter_cache" do
     context "downloaded" do
       let!(:scorecard) { create(:scorecard, downloaded_count: 0) }
